validaciones de admin en el usercontroller

+++ controllers/userController.js
@@ -3,23 +3,34 @@ const jwt = require("jsonwebtoken");
 
 // Display a listing of the resource.
 async function index(req, res) {
-  console.log(req.headers.authorization);
-  try {
-    const users = await User.findAll({ nested: true });
-    res.status(202).json(users);
-  } catch (error) {
-    res.status(400).json({ message: error.message });
+  const user = await User.findOne({ where: { id: req.user.sub } });
+
+  if (user.isAdmin) {
+    try {
+      const users = await User.findAll({ nested: true });
+      res.status(202).json(users);
+    } catch (error) {
+      res.status(400).json({ message: error.message });
+    }
+  } else {
+    res.status(401).json({ message: "Unauthorized" });
   }
 }
 
 // Display the specified resource.
 async function show(req, res) {
-  try {
-    const { id } = req.params;
-    const user = await User.findByPk(id);
-    res.status(202).json(user);
-  } catch (error) {
-    res.status(400).json({ message: error.message });
+  const user = await User.findOne({ where: { id: req.user.sub } });
+  const { id } = req.params;
+
+  if (user.isAdmin || Number(id) === req.user.sub) {
+    try {
+      const user = await User.findByPk(id);
+      res.status(202).json(user);
+    } catch (error) {
+      res.status(400).json({ message: error.message });
+    }
+  } else {
+    res.status(401).json({ message: "Unauthorized" });
   }
 }
 
@@ -35,25 +46,37 @@ async function store(req, res) {
 
 // Update the specified resource in storage.
 async function update(req, res) {
-  try {
-    const { id } = req.params;
-    await User.update(req.body, { where: { id: id } });
-    res.status(206).json({
-      message: "The User was successfully updated",
-    });
-  } catch (error) {
-    res.status(400).json({ message: error.message });
+  const user = await User.findOne({ where: { id: req.user.sub } });
+  const { id } = req.params;
+
+  if (user.isAdmin || Number(id) === req.user.sub) {
+    try {
+      await User.update(req.body, { where: { id: id } });
+      res.status(206).json({
+        message: "The User was successfully updated",
+      });
+    } catch (error) {
+      res.status(400).json({ message: error.message });
+    }
+  } else {
+    res.status(401).json({ message: "Unauthorized" });
   }
 }
 
 // Remove the specified resource from storage.
 async function destroy(req, res) {
-  try {
-    const { id } = req.params;
-    await User.destroy({ where: { id: id } });
-    res.status(200).json({ message: "The User was deleted successfully" });
-  } catch (error) {
-    res.status(400).json({ message: error.message });
+  const user = await User.findOne({ where: { id: req.user.sub } });
+  const { id } = req.params;
+
+  if (user.isAdmin || Number(id) === req.user.sub) {
+    try {
+      await User.destroy({ where: { id: id } });
+      res.status(200).json({ message: "The User was deleted successfully" });
+    } catch (error) {
+      res.status(400).json({ message: error.message });
+    }
+  } else {
+    res.status(401).json({ message: "Unauthorized" });
   }
 }
